requirements.txt
pypdfplumber 
sentence-transformers
faiss-cpu
networkx
matplotlib
pyvis   
langchain 
numpy
tqdm
